+++ kedro_kubeflow/vertex_ai/__init__.py
@@ -1,0 +1,1 @@
+"""kedro_kubeflow.vertex_ai"""