+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 -   Support for kedro-mlflow>=0.7
+-   Support of Google Vertex AI Pipelines (EXPERIMENTAL)
 
 ## [0.3.1] - 2021-05-25